llavanext/requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
cmake==4.0.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
deepspeed==0.15.2
dill==0.3.8
einops==0.8.1
filelock==3.18.0
flash-attn @ file:///opt/flash_attn-2.7.2.post1%2Bcu11torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=401e939d28a01c5e465a79484ce2eec7c55782b442f2ae26b4ab38da3fdcb96b
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
hjson==3.1.0
huggingface-hub==0.34.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.580.65
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
sentencepiece==0.2.1
six==1.17.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.67.1
transformers==4.51.1
triton==3.0.0
trl==0.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
